Fix workspace/xreferences
